requirements.txt
Pillow==9.3.0
PyGithub==1.57
python-dotenv==0.21.0
requests==2.28.1
